Allow docker container to run in non-interactive mode
bash exits if the container isn't started in non-interactive mode, sleep
won't (at least for 365 days, in this case). The CMD can still be
overridden to start bash if the user desires.

Using sleep is a bit of a hack, and really should be replaced with
something that deals with signals correctly (i.e. allows stopping the
container without killing it).