Switch pipe to a streaming approach, rather than microbatching
A more efficient (faster) approach. Leaves it largely up to the consumer
and producer to work out how big batches from/to Kafka should be.

Requires taking over committing from the consumer, however, as we don't
want to commit a message while the produce it generated has yet to be
acknowledged.